+++ dataworkspace/dataworkspace/apps/api_v1/urls.py
@@ -59,4 +59,11 @@ urlpatterns = [
             ('dataworkspace.apps.api_v1.accounts.urls', 'account'), namespace='account'
         ),
     ),
+    path(
+        'application-instance/',
+        include(
+            ('dataworkspace.apps.api_v1.applications.urls', 'application-instance'),
+            namespace='application-instance',
+        ),
+    ),
 ]