+++ dataworkspace/dataworkspace/tests/factories.py
@@ -1,5 +1,5 @@
 import uuid
-from datetime import datetime
+from datetime import datetime, timedelta
 
 import factory.fuzzy
 
@@ -223,6 +223,7 @@ class ApplicationTemplateFactory(factory.django.DjangoModelFactory):
         model = 'applications.ApplicationTemplate'
 
 
+<<<<<<< HEAD
 class ApplicationTemplateUserPermissionFactory(factory.django.DjangoModelFactory):
     user = factory.SubFactory(UserFactory)
     application_template = factory.SubFactory(ApplicationTemplateFactory)
@@ -256,3 +257,27 @@ class VisualisationCatalogueItemFactory(factory.django.DjangoModelFactory):
 
     class Meta:
         model = 'datasets.VisualisationCatalogueItem'
+
+
+class ApplicationInstanceReportFactory(factory.django.DjangoModelFactory):
+    id = factory.LazyAttribute(lambda _: uuid.uuid4())
+    application_template = factory.SubFactory(ApplicationTemplateFactory)
+    owner = factory.SubFactory(UserFactory)
+    public_host = 'https://analysisworkspace.dev.uktrade.io/'
+    spawner = factory.fuzzy.FuzzyText()
+    spawner_application_template_options = '{}'
+    spawner_application_instance_id = factory.LazyAttribute(lambda _: uuid.uuid4())
+    spawner_created_at = datetime.now() - timedelta(minutes=5)
+    spawner_stopped_at = datetime.now()
+    spawner_cpu = factory.fuzzy.FuzzyChoice(['256', '1024', '2048', '4096'])
+    spawner_memory = factory.fuzzy.FuzzyChoice(['512', '8192', '16384', '30720'])
+    state = factory.fuzzy.FuzzyChoice(['SPAWNING', 'RUNNING', 'STOPPED'])
+    proxy_url = 'https://analysisworkspace.dev.uktrade.io/'
+    cpu = factory.fuzzy.FuzzyChoice(['256', '1024', '2048', '4096'])
+    memory = factory.fuzzy.FuzzyChoice(['512', '8192', '16384', '30720'])
+    single_running_or_spawning_integrity = factory.fuzzy.FuzzyText()
+    commit_id = factory.fuzzy.FuzzyText(length=8)
+
+    class Meta:
+        model = 'applications.ApplicationInstanceReport'
+